requirements.txt
absl-py==2.3.1
accelerate==1.8.1
addict==2.4.0
aiofiles==22.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.21.0
altair==4.2.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-timeout==5.0.1
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
billiard==4.2.1
bleach==6.2.0
blessed==1.21.0
blinker==1.9.0
boto3==1.39.4
botocore==1.39.4
cachetools==6.1.0
celery==5.5.3
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1752493286596/work/certifi
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cmake==4.0.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7.1
configer @ git+https://github.com/MPI-IS/configer.git@8cd1e3e556d9697298907800a743e120be57ac36
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.0
croniter==1.4.1
cryptography==45.0.5
cycler==0.12.1
Cython==3.1.2
dash==3.1.1
dateutils==0.6.12
debugpy==1.8.14
decorator==5.2.1
deepdiff==7.0.1
defusedxml==0.7.1
deprecation==2.1.0
dictdiffer==0.9.0
diff_gaussian_rasterization @ file:///mnt/beegfs/home/stud220/git/ImageTo3DSegmentedClothes/Human3Diffusion/gaussian-opacity-fields/submodules/diff-gaussian-rasterization
diffusers==0.34.0
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
docopt==0.6.2
docstring_parser==0.16
dotmap==1.3.30
dpath==2.2.0
dulwich==0.23.2
dvc==3.61.0
dvc-data==3.16.10
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
editor==1.6.6
einops==0.8.1
entrypoints==0.4
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.1.1
flatbuffers==25.2.10
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.58.5
fqdn==1.5.1
freetype-py==2.5.1
frozenlist==1.7.0
fsspec==2024.6.1
funcy==2.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
grandalf==0.8
grpcio==1.73.1
gto==1.7.2
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.33.2
humanfriendly==10.0
hydra-core==1.2.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inquirer==3.4.0
iopath==0.1.10
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.7
isoduration==20.11.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.5.1
Js2Py==0.74
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.8.1
jupyter_packaging==0.12.3
jupyter_server==2.16.0
jupyter_server_fileid==0.9.3
jupyter_server_terminals==0.5.3
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiui==0.2.16
kiwisolver==1.4.7
kombu==5.5.4
kornia==0.8.1
kornia_rs==0.1.9
lazy_loader==0.4
lightning==2.0.9.post0
lightning-cloud==0.5.70
lightning-utilities==0.14.3
llvmlite==0.43.0
loguru==0.7.3
-e git+https://github.com/riccardomarin/NICP@93147a8108210ef95adebd2c13afaaaf34d7eb25#egg=lvd_templ
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
multidict==6.6.3
narwhals==1.45.0
nbclassic==1.3.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nn-template-core==0.1.1
notebook==6.5.7
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
objprint==0.3.0
omegaconf==2.3.0
onnxruntime==1.19.2
open3d==0.15.2
opencv-python==4.6.0.66
opencv-python-headless==4.11.0.86
ordered-set==4.1.0
orjson==3.10.18
overrides==7.7.0
oyaml==1.0
packaging==24.2
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pathtools==0.1.2
pexpect==4.9.0
pillow==11.3.0
pipwin==0.5.2
platformdirs==4.3.8
plotly==5.10.0
plyfile==1.1.2
pooch==1.8.2
portalocker==3.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==3.20.3
psbody-mesh @ file:///mnt/beegfs/home/stud220/git/mesh
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==25.5.0
pyarrow==20.0.0
pycparser==2.22
pycuda==2025.1.1
pydantic==2.1.1
pydantic_core==2.4.0
pydeck==0.9.1
pyDeprecate==0.3.1
pydot==4.0.1
pygit2==1.15.1
pyglet==2.1.6
Pygments==2.19.2
pygtrie==2.5.0
pyjsparser==2.7.1
PyJWT==2.10.1
PyMatting==1.1.14
PyMCubes==0.1.6
Pympler==1.1
PyOpenGL==3.1.0
pyparsing==3.2.3
PyPrind==2.11.3
pyquaternion==0.9.9
pyrender==0.1.45
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytools==2024.1.14
pytorch-lightning==1.5.10.post0
-e git+https://github.com/facebookresearch/pytorch3d.git@3aee2a6005cbf32275d705b4440deb3813f6a71e#egg=pytorch3d
pytz==2025.2
PyWavelets==1.6.0
PyYAML==6.0.2
pyzmq==27.0.0
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
rembg==2.0.61
requests==2.32.4
retrying==1.4.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
robust-laplacian==0.2.4
rpds-py==0.26.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
runs==1.2.2
s3transfer==0.13.0
safetensors==0.5.3
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.13.1
scmrepo==3.3.11
semver==3.0.4
Send2Trash==1.8.3
sentry-sdk==2.32.0
setproctitle==1.3.6
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
simple_knn @ file:///mnt/beegfs/home/stud220/git/ImageTo3DSegmentedClothes/Human3Diffusion/gaussian-opacity-fields/submodules/simple-knn
six==1.17.0
smmap==5.0.2
smplx==0.1.28
sniffio==1.3.1
soupsieve==2.7
sqltrie==0.11.2
stack-data==0.6.3
starlette==0.47.1
starsessions==1.3.0
stqdm==0.0.5
streamlit==1.22.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2024.8.30
tinycss2==1.4.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
toolz==1.0.0
torch==2.2.0+cu121
torchaudio==2.2.0+cu121
torchmetrics==1.7.4
torchvision==0.17.0+cu121
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.1
transforms3d==0.3.1
trimesh==3.13.0
triton==2.2.0
typeguard==4.4.4
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
tyro==0.9.26
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.35.0
validators==0.35.0
varname==0.15.0
vine==5.1.0
voluptuous==0.15.2
wandb==0.15.2
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xatlas==0.0.11
xformers==0.0.24
xmod==1.8.1
y-py==0.6.2
yarl==1.20.1
ypy-websocket==0.8.4
zc.lockfile==3.0.post1
zipp==3.23.0
